refactor: cleanup empty topics properly
The implementation for this is rather ugly, due to the annoying
behaviour of `select` fusing its streams, meaning `poll_next` is never
called again on the `FuturesUnordered` once it's empty. To work around
this, we can split the `select` and recombine it, which feels like a
fairly expensive way to reset the `FusedStream`.